% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cumulative.R
\name{add_cumulative}
\alias{add_cumulative}
\title{Add a cumulative sum column to a data.frame.}
\usage{
add_cumulative(dat, colname, dir = "down")
}
\arguments{
\item{dat}{data.frame to add cumulative sum to}

\item{colname}{specify the unquoted name of the column to sum, or leave it blank in which case the function will default first to a column called "valid_percent" and then for "percent".  These defaults support running this function on the result of calls to \code{janitior::tabyl}.  If no colname is supplied and these default columns are absent in the data.frame, the function will error.}

\item{dir}{direction to sum; defaults to "down" but can sum from the bottom of a data.frame with "up".  In this case the resulting column name will be "cumulative_up".}
}
\value{
a data.frame.
}
\description{
A tidyverse-style function to add a cumsum column.
}
\examples{
library(janitor)
mtcars \%>\%
  add_cumulative(mpg)

mtcars \%>\%
  tabyl(cyl) \%>\%
  add_cumulative()

# Vector with an NA
x <- c(0, 1, 2, 3, 3, 3, NA)

x \%>\%
  tabyl() \%>\%
  add_cumulative()

x \%>\%
  tabyl() \%>\%
  add_cumulative(dir = "up")
}
