% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart_counts.R
\name{bar_chart_counts}
\alias{bar_chart_counts}
\title{Bar chart of counts with TNTP polish}
\usage{
bar_chart_counts(df = NULL, var, group_var, display = "n", group_colors,
  title = NULL, var_label, var_color = palette_tntp("medium_blue"),
  digits = 1, font = "Segoe UI", font_size = 12)
}
\arguments{
\item{df}{the data.frame to be used in the bar chart}

\item{var}{unquoted column name for desired variable}

\item{group_var}{(optional) unquoted column name for group variable}

\item{display}{bar chart type "n" by default; also accepts "pct"}

\item{group_colors}{(optional) character vector of group colors}

\item{title}{main chart title}

\item{var_label}{label for x-axis}

\item{var_color}{color for non-grouped charts; set to dark_blue by default}

\item{digits}{integer indicating the number of decimal places to be used}

\item{font}{font for chart text; Segoe UI by default}

\item{font_size}{size for chart text; set to 12 by default}
}
\description{
Takes a user supplied data frame and turns the designated column into
an N bar chart (uses position dodge from ggplot2).
}
\examples{

# An  bar chart by default
mtcars \%>\%
 bar_chart_counts(var     = cyl,
                title   = "Number of mtcars by cylinder")

# With a grouping variable
mtcars \%>\%
 bar_chart_counts(var          = cyl,
                group_var    = vs,
                title        = "Percentage of mtcars by cylinder")

# with colors
bar_chart_counts(mtcars, am, cyl, group_colors = palette_tntp("orange", "green", "dark_blue"), display = "pct")
}

