% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart_tntp.R
\name{bar_chart_tntp}
\alias{bar_chart_tntp}
\title{Bar chart with TNTP polish}
\usage{
bar_chart_tntp(df = NULL, var = NULL, display = "n", title = NULL, ...)
}
\arguments{
\item{df}{the data.frame to be used in the bar chart}

\item{var}{unquoted column name for desired variable}

\item{display}{set to "n" by default; will also accept "percent"}

\item{title}{main chart title}

\item{...}{additional arguments}
}
\description{
Takes a user supplied data frame and turns the designated column into
a bar chart. Can specify an "n" bar chart or a "percent" bar chart.
}
\examples{

# An N bar chart by default
mtcars \%>\%
 bar_chart_tntp(var     = cyl,
                title   = "Number of mtcars by cylinder")

#'mtcars \%>\%
 bar_chart_tntp(var     = cyl,
                title   = "Percentage of mtcars by cylinder",
                display = "percent")
}

