% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_utils.R
\name{check_all_recode}
\alias{check_all_recode}
\title{Process a range of check-all-that-apply response columns for correct tabulation.}
\usage{
check_all_recode(dat, ...)
}
\arguments{
\item{dat}{a data.frame with survey data}

\item{...}{unquoted variable names containing the answer choices.  Can be specified as a range, i.e., \code{q1_1:q1_5} or using other helper functions from \code{dplyr::select()}.}
}
\value{
Returns the original data.frame with the selected column range updated.
}
\description{
Some survey software returns check-all-that-apply response columns where missing values could indicate either that the respondent skipped the question entirely, or that they did not select that particular answer choice.  To count the responses properly, the cases where a respondent did not check any of choices - i.e., they skipped the question - should not be counted in the denominator (assuming that the choices were completely exhaustive, or that there was an NA option).

This function takes a data.frame and range of columns containing all answer choices to a check-all-that-apply question and updates the columns in the data.frame to contain one of three values: the original value if the choice was selected; "did not select" if the respondent chose another option but not this one; or NA if the respondent skipped the question (i.e., they did not select any of the choices) and thus their response is truly missing.

\code{check_all_recode()} prepares the data.frame for a call to its sister function \code{check_all_count()}.
}
\examples{
x <- data.frame( # 4th person didn't respond at all
  unrelated = 1:4,
  q1_1 = c("a", "a", "a", NA),
  q1_2 = c("b", "b", NA, NA),
  q1_3 = c(NA, NA, "c", NA)
)
library(dplyr) # for the \%>\% pipe
x \%>\%
  check_all_recode(q1_1:q1_3)

# You can use any of the dplyr::select() helpers to identify the columns:
x \%>\%
  check_all_recode(contains("q1"))

}
