% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_tntp.R
\docType{data}
\name{colors_tntp}
\alias{colors_tntp}
\title{TNTP colors}
\format{
An object of class \code{character} of length 34.
}
\usage{
colors_tntp
}
\description{
This list of colors has been superseded by the new brand colors and the new
function \code{\link[=tntp_colors]{tntp_colors()}}.
}
\examples{
\dontrun{
tntp_colors()
}
}
\keyword{datasets}
