% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_tntp.R
\docType{data}
\name{colors_tntp_classic}
\alias{colors_tntp_classic}
\title{Classic pallette}
\format{
An object of class \code{character} of length 14.
}
\usage{
colors_tntp_classic
}
\description{
Classic pallette
}
\examples{
\dontrun{
 colors_tntp_classic
}
}
\keyword{datasets}
