% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_tntp.R
\docType{data}
\name{colors_tntp_default}
\alias{colors_tntp_default}
\title{Default pallette}
\format{
An object of class \code{character} of length 12.
}
\usage{
colors_tntp_default
}
\description{
Default pallette
}
\examples{
\dontrun{
 colors_tntp_default
}
}
\keyword{datasets}
