% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_tntp.R
\docType{data}
\name{colors_tntp_palette}
\alias{colors_tntp_palette}
\title{TNTP pallette}
\format{
An object of class \code{character} of length 16.
}
\usage{
colors_tntp_palette
}
\description{
TNTP pallette
}
\examples{
\dontrun{
colors_tntp_palette
}
}
\keyword{datasets}
