% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_agreement.R
\name{factorize_agreement_df}
\alias{factorize_agreement_df}
\title{Convert all agreement-seeming character vectors in a data.frame to factors, calling \code{factorize_agreement()} on each of those.}
\usage{
factorize_agreement_df(dat, lvls, cutoff = 1)
}
\arguments{
\item{dat}{data.frame with some agreement variables stored as characters.}

\item{lvls}{(optional) the agreement levels in your survey, in order.  If you have different agreement levels on the same survey, don't use this function, and use \code{factorize_agreement()} on the individual columns.  Default is a 6-pt scale from Strongly Disagree to Strongly Agree.}

\item{cutoff}{(optional) what is the threshold for % of text values matching the agreement levels that causes factorization?  If unsure, leave at 1 - it should probably be 1 permanently with better error catching of values that don't fit the stated values paradigm.}
}
\value{
data.frame with factorization completed in place.
}
\description{
Convert all agreement-seeming character vectors in a data.frame to factors, calling \code{factorize_agreement()} on each of those.
}
\examples{
library(dplyr)
mtcars \%>\%
  dplyr::mutate(agr = rep(c("Somewhat agree", "Strongly Disagree"), 16)) \%>\%
  factorize_agreement_df()
}
