% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_agreement.R
\name{factorize_df}
\alias{factorize_df}
\title{Convert all character vectors containing a set of values in a data.frame to factors.}
\usage{
factorize_df(dat, lvls, cutoff = 1)
}
\arguments{
\item{dat}{data.frame with some factor variables stored as characters.}

\item{lvls}{The factor levels in your variable(s), in order.  If you have a question whose possible responses are a subset of another question's, don't use this function; manipulate the specific columns with \code{dplyr::mutate_at}.}

\item{cutoff}{(optional) what is the threshold for % of text values matching the agreement levels that causes factorization?  If unsure, leave at 1 - it should probably be 1 permanently with better error catching of values that don't fit the stated values paradigm.}
}
\value{
data.frame with factorization completed in place.
}
\description{
Convert all character vectors containing a set of values in a data.frame to factors.
}
\examples{
library(dplyr)
mtcars \%>\%
  dplyr::mutate(agr = rep(c("Somewhat Agree", "Strongly disagree"), 16)) \%>\%
  factorize_df(lvls = c("Strongly disagree", "Somewhat disagree", "Somewhat agree", "Strongly agree"))

# prints warning due to case mismatches:
mtcars \%>\%
  dplyr::mutate(agr = rep(c("Somewhat Agree", "Strongly Disagree"), 16)) \%>\%
  factorize_df(lvls = c("Strongly disagree", "Somewhat disagree", "Somewhat agree", "Strongly agree"))
}
