% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_lump_ends.R
\name{fct_lump_ends}
\alias{fct_lump_ends}
\title{Collapse factor levels into three groups: top-n levels, bottom-n levels, and all other levels.}
\usage{
fct_lump_ends(input_vec, n = 2)
}
\arguments{
\item{input_vec}{the factor variable to lump.}

\item{n}{number of levels to include in top and bottom groups}
}
\value{
An ordered factor.
}
\description{
Typical usage is for say, a Likert question on a survey where we group respondents into strongly disagree/disagree, somewhat disagree/somewhat agree, and agree/strongly agree.
}
\examples{
fct_lump_ends(as.factor(mtcars$hp), 3)
}
