% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-utils.R
\name{figureN}
\alias{figureN}
\title{Create sequential figure numbers}
\usage{
figureN(x)
}
\arguments{
\item{x}{character string description of the figure}
}
\value{
An atomic character vector prepended with a Figure number
}
\description{
Create sequential figure numbers
}
\examples{

figureN("Distribution of cars by cylinder count")
# Inline RMarkdown code: `r figureN("Distribution of cars by cylinder count")`

#
}
