% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_.R
\name{get_ext}
\alias{get_ext}
\title{Pull extension from a path}
\usage{
get_ext(path)
}
\arguments{
\item{path}{file path}
}
\value{
The extension as a character vector
}
\description{
Pull extension from a path
}
