% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntp_style.R
\name{get_usable_family}
\alias{get_usable_family}
\title{Checks if a font family is usable and returns a usable font if not}
\usage{
get_usable_family(family, silent = FALSE, default_family = "sans")
}
\arguments{
\item{family}{the font family to check as a character}

\item{silent}{logical. If TRUE doesn't raise a warning if the font family is unavailable}

\item{default_family}{defaults to "sans", but can be set to another fallback family.}
}
\value{
a character of a usable font family
}
\description{
Helper function. Checks if a given family value is available, and if not returns
the default font family ("sans" or user provided)
}
