% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_tntp.R
\name{header_tntp}
\alias{header_tntp}
\title{Insert header_script_tntp.}
\usage{
header_tntp()
}
\value{
nothing
}
\description{
Call this function from inside a .R file in RStudio to insert the standard
TNTP header into your active script.
}
\examples{
\dontshow{if (rstudioapi::isAvailable()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
header_tntp()
\dontshow{\}) # examplesIf}
}
