% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tntp_2018.R
\name{import_segoe_ui}
\alias{import_segoe_ui}
\title{Import Segoe UI Condensed font for use in charts}
\usage{
import_segoe_ui()
}
\value{
nothing
}
\description{
This function will check if Segoe UI is already accessible in R and if not
it will attempt to import it using the \code{extrafont} package
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
import_segoe_ui()
\dontshow{\}) # examplesIf}
}
