% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_.R
\name{is_drive_id}
\alias{is_drive_id}
\title{Test if a string is formatted like a Sharepoint drive id}
\usage{
is_drive_id(x)
}
\arguments{
\item{x}{string to test}
}
\value{
TRUE or FALSE
}
\description{
Test if a string is formatted like a Sharepoint drive id
}
