% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_.R
\name{is_site_id}
\alias{is_site_id}
\title{Test if a string is formatted like a Sharepoint site id}
\usage{
is_site_id(x)
}
\arguments{
\item{x}{string to test}
}
\value{
TRUE or FALSE
}
\description{
Test if a string is formatted like a Sharepoint site id
}
