% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_.R
\name{is_site_url}
\alias{is_site_url}
\title{Test if a string is formatted like a Sharepoint site url}
\usage{
is_site_url(x)
}
\arguments{
\item{x}{string to test}
}
\value{
TRUE or FALSE
}
\description{
Test if a string is formatted like a Sharepoint site url
}
