% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_tntp.R
\docType{data}
\name{palette_names}
\alias{palette_names}
\title{Pallette names}
\format{
An object of class \code{character} of length 7.
}
\usage{
palette_names
}
\description{
This list of palette names has been superseded by the new brand colors and
new functions \code{\link[=tntp_colors]{tntp_colors()}} and \code{\link[=tntp_palette]{tntp_palette()}}. To see all of the new
brand palettes, use \code{\link[=show_tntp_palette]{show_tntp_palette()}}.
}
\examples{
\dontrun{
show_tntp_palette()
}
}
\keyword{datasets}
