% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_tntp.R
\name{palette_tntp_scales}
\alias{palette_tntp_scales}
\title{scale_palette_tntp}
\usage{
palette_tntp_scales(palette = c("default", "likert_4pt", "likert_5pt",
  "likert_6pt", "colors_tntp_classic"))
}
\arguments{
\item{palette}{}
}
\description{
scale_palette_tntp
}
