% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_sy.R
\name{parse_date}
\alias{parse_date}
\title{Attempt to parse a date with common formats}
\usage{
parse_date(date)
}
\arguments{
\item{date}{a character or Date vector to parse}
}
\value{
Date vector
}
\description{
Helper function for date_to_sy. Returns a date object as is, or noisily attempts to parse
a string in the form YYYY-MM-DD or MM/DD/YYYY. If the date cannot be parsed,
throws an error.
}
