% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_diverge.R
\name{position_diverge}
\alias{position_diverge}
\title{Easy Diverging Bar Charts}
\usage{
position_diverge(vjust = 1, break_after = NULL)
}
\arguments{
\item{vjust}{OPTIONAL set to 0.5 for geom_text() or geom_label() to center within the bar.}

\item{break_after}{OPTIONAL. Either an integer index or character value that is the last positive level}
}
\description{
This is a modification of `ggplot2::position_stack()` for creating diverging bar charts.
In order to use this function, you must set a fill aesthetic and that aesthetic must be
a factor. This function will automatically break your chart into negative and positive
values and display them in the same order as geom_col()
}
