% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_.R
\name{process_type}
\alias{process_type}
\title{Determine function type by extension and provided type
Handles type validation}
\usage{
process_type(ext, type)
}
\arguments{
\item{ext}{File extension (from get_ext)}

\item{type}{User-specified type}
}
\value{
a type ("dataframe" "rds" or "rdata")
}
\description{
Determine function type by extension and provided type
Handles type validation
}
