% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_agreement.R
\name{prop_agreement}
\alias{prop_agreement}
\title{Calculate the percent of non-missing values in a character vector containing the term "agree."  This is a helper function for factorize_agreement_df().}
\usage{
prop_agreement(vec)
}
\arguments{
\item{vec}{character vector.}
}
\value{
numeric proportion between 0 and 1.
}
\description{
Calculate the percent of non-missing values in a character vector containing the term "agree."  This is a helper function for factorize_agreement_df().
}
