% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_utils.R
\name{recode_top_2}
\alias{recode_top_2}
\title{Recode a variable into "Top-2" or "Not in Top-2".}
\usage{
recode_top_2(x, top_2_values = c("strongly agree", "agree"))
}
\arguments{
\item{x}{the character or factor vector to be recoded}

\item{top_2_values}{a character vector with the strings that should be considered "Top-2".  Defaults to "strongly agree" and "agree" but can be overwritten.}
}
\value{
a factor variable (for nicer ordering in calls to \code{janitor::tabyl}) with values mapped to Top-2 and Not in Top-2.
}
\description{
Recodes a character variable into a binary result: "Top-2" if the value matches any of the supplied character strings in the \code{top_2_values} vector, "Not in Top-2" if not.  NA remains NA.

This recoding is not case-sensitive; if you specify "agree" as a top-2 value, "Agree" will be counted as Top-2, and vice versa.

You don't *have* to specify two top levels; you can specify one, three, bottom levels, etc.
#'
}
\examples{
agreement <- c("Strongly agree", "Agree", "Somewhat agree", "Somewhat disagree", "Strongly disagree", "Frogs", NA)
recode_top_2(agreement) # default values of "strongly agree" and "agree" are used for recoding
recode_top_2(agreement, "frogs")

freq <- c("always", "often", "sometimes", "never")
recode_top_2(freq, "always")
}
