% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_tntp.R
\name{scale_fill_tntp}
\alias{scale_fill_tntp}
\title{scale_fill_tntp}
\usage{
scale_fill_tntp(palette = palette_names, drop = FALSE, ...)
}
\arguments{
\item{palette}{One of: "default", "likert_4pt", "likert_5pt", "likert_6pt", "colors_tntp_old"}

\item{...}{}
}
\description{
scale_fill_tntp
}
