% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-utils.R
\name{set_data_memo_formatting}
\alias{set_data_memo_formatting}
\title{Set the formatting options for a TNTP Data Memo}
\usage{
set_data_memo_formatting()
}
\value{
nothing
}
\description{
internal function that calls standard formatting options for the Data Memo RMarkdown template
moved here to keep the actual memo template cleaner and easier to use
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
set_data_memo_formatting()
\dontshow{\}) # examplesIf}
}
