% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_in_excel.R
\name{show_in_excel}
\alias{show_in_excel}
\title{Write Dataframe to a temp excel file and open it.}
\usage{
show_in_excel(.data)
}
\arguments{
\item{.data}{Dataframe}
}
\value{
nothing
}
\description{
Write Dataframe to a temp excel file and open it.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# View a data set in excel
mtcars |> show_in_excel()

\dontshow{\}) # examplesIf}
}
