% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tntp_colors.R
\name{show_tntp_colors}
\alias{show_tntp_colors}
\title{Show TNTP color palette}
\usage{
show_tntp_colors(palette = "all")
}
\arguments{
\item{palette}{Default is 'all' tntp colors, but can also select a specific
palette by passing either 'likert' or 'likert_orange_to_green}
}
\value{

}
\description{
A quick and dirty way to show tntp colors in a plot.
}
\examples{
show_tntp_colors(palette = "likert")
}
