% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymonkey.R
\name{sm_get_responses}
\alias{sm_get_responses}
\title{Get information about respondents to a survey}
\usage{
sm_get_responses(auth_token, survey_id, page = 1, per_page = 250)
}
\arguments{
\item{auth_token}{Get this from your SurveyMonkey app}

\item{survey_id}{The ID of the survey you want information about; use \code{sm_get_survey()} to obtain this.}

\item{page}{Which page of results to to display?}

\item{per_page}{How many results to display?}
}
\description{
Call on a single survey ID retrieved using \code{sm_get_surveys()}, to get a list with information about the respondents to a particular survey.
}

