% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymonkey.R
\name{sm_get_surveys}
\alias{sm_get_surveys}
\title{Get the list of surveys from a SurveyMonkey account}
\usage{
sm_get_surveys(auth_token, page = 1, per_page = 250)
}
\arguments{
\item{auth_token}{Get this from your SurveyMonkey app}

\item{page}{Which page of results to display?}

\item{per_page}{How many results to display?}
}
\value{
a list with information about the surveys in the account.  You probably want to access $data, the data.frame with survey IDs, for use in other functions.
}
\description{
Returns information about the surveys in a SurveyMonkey account.
}

