% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_.R
\name{sp_check_folder}
\alias{sp_check_folder}
\title{Check for existence of a Sharepoint folder and offer to create it if it
doesn't exist}
\usage{
sp_check_folder(site, drive, folder_path)
}
\arguments{
\item{site}{ms_site object}

\item{drive}{ms_drive object}

\item{folder_path}{path to an item}
}
\value{
nothing
}
\description{
Check for existence of a Sharepoint folder and offer to create it if it
doesn't exist
}
