% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_.R
\name{sp_read_xlsx}
\alias{sp_read_xlsx}
\title{Internal function for reading excel files from Sharepoint}
\usage{
sp_read_xlsx(path, drive, ...)
}
\arguments{
\item{path}{path}

\item{drive}{ms_drive object}

\item{...}{additional arguments from sp_read()}
}
\value{
data read by readxl::read_excel()
}
\description{
Internal function for reading excel files from Sharepoint
}
