% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_.R
\name{sp_site}
\alias{sp_site}
\alias{sp_drive}
\title{Return Microsoft365R site or drive object}
\usage{
sp_site(site)

sp_drive(site, drive)
}
\arguments{
\item{site}{Site identifier. Can be the site_name, site_url or site_id. If not provided, returns the stored default site if it exists.}

\item{drive}{Name of the drive within the site. If site is provided but drive is not, uses the first drive of the provided site. If neither is provided, uses the stored default drive if it exists.}
}
\value{
A Microsoft365R site object or drive object
}
\description{
Pulls the site or drive (if given) or returns the stored default
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Get current default site
x <- sp_site()

# Get specified site
x <- sp_site("Data Analytics")
\dontshow{\}) # examplesIf}
}
