% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_.R
\name{sp_string}
\alias{sp_string}
\title{Internal function to create cli-friendly site/drive/path string}
\usage{
sp_string(
  site = NULL,
  site_name = NULL,
  drive = NULL,
  drive_name = NULL,
  path = NULL
)
}
\arguments{
\item{site}{site object (from sp_site())}

\item{site_name}{site name}

\item{drive}{drive object (from sp_drive())}

\item{drive_name}{drive name}

\item{path}{path}
}
\value{
a cli-formatted string
}
\description{
Can use either a site_name or site object, a drive_name or drive_object
}
