% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_.R
\name{sp_write_xlsx}
\alias{sp_write_xlsx}
\title{Function for uploading an xls / xlsx file}
\usage{
sp_write_xlsx(x, path, drive, ...)
}
\arguments{
\item{x}{R object}

\item{path}{path on the Sharepoint drive}

\item{drive}{ms_drive object}

\item{...}{additional arguments passed on from sp_write()}
}
\value{
nothing
}
\description{
Function for uploading an xls / xlsx file
}
