% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize.R
\name{standardize_case}
\alias{standardize_case}
\title{Update case of a character vector}
\usage{
standardize_case(vec, new_case)
}
\arguments{
\item{vec}{The character vector you want to update}

\item{new_case}{A character vector of correctly cased strings}
}
\value{
a character vector the same length as vec
}
\description{
Helper function for factorize_df() and safe_factor(). Returns a vector of
the same length as vec, with any values that match values in valid_strings
updated to the case in valid_strings
}
