% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-utils.R
\name{tableN}
\alias{tableN}
\title{Create sequential table numbers}
\usage{
tableN(x)
}
\value{
nothing
}
\description{
Create sequential table numbers
}
\examples{
# not run, in RMarkdown doc: `r tableN("Distribution of cars by cylinder count")`
}

