% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_utils.R
\name{tabulate_check_all}
\alias{tabulate_check_all}
\title{Tabulate a range of check-all-that-apply response columns in a single table.}
\usage{
tabulate_check_all(dat, ...)
}
\arguments{
\item{dat}{a data.frame with survey data}

\item{...}{unquoted variable names containing the answer choices.  Can be specified as a range, i.e., \code{q1_1:q1_5} or using other helper functions from \code{dplyr::select()}.}
}
\value{
Returns a data.frame with the tabulated results (n and % of question respondents choosing each choice.)
}
\description{
This function is to be run on columns treated with \code{treat_check_alls()}.

Takes a data.frame and range of columns containing all answer choices to a check-all-that-apply question and tabulates the results.  People who did not select any choices  (i.e., they did not answer the question) are omitted from the denominator.  For this to make sense, the question's choices should be MECE, or there should be an NA option.
}
\examples{
x <- data.frame( # 4th person didn't respond at all
  unrelated = 1:4,
  q1_1 = c("a", "a", "a", NA),
  q1_2 = c("b", "b", NA, NA),
  q1_3 = c(NA, NA, "c", NA)
)
library(dplyr) # for the \%>\% pipe
x \%>\%
  treat_check_alls(q1_1:q1_3) \%>\%
  tabulate_check_all(q1_1:q1_3)

}
