% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teacher_survey}
\alias{teacher_survey}
\title{Teacher survey data}
\format{
## `teacher survey`
A data frame with 5 columns and 20 rows. The five columns are a `timing` column,
followed by four column for each of the four high expectations questions. Responses
are on the 'strongly agree' to strongly disagree' 6-point scale.
}
\source{
simulated in `data-raw/teacher_survey.R`
}
\usage{
teacher_survey
}
\description{
Simulated teacher survey data. Data only includes teh four TNTP high expectations questions.
}
\keyword{datasets}
