% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tntp.R
\name{theme_tntp}
\alias{theme_tntp}
\title{TNTP's ggplot2 theme}
\usage{
theme_tntp(show_legend_title = TRUE, base_size = 12,
  base_family = "Segoe UI", grid_color = "grey93")
}
\arguments{
\item{show_legend_title}{logical. Should the legend title be shown?  Leave as \code{TRUE} if you want to change the legend title with a subsequent line \code{+ labs(...)}.}

\item{base_size}{base font size}

\item{base_family}{base font family}

\item{grid_color}{color for major gridlines}
}
\description{
ggplot2 theme customized for TNTP aesthetics
}
\examples{
library(ggplot2)
library(extrafont)

p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg,
     colour=factor(gear))) + facet_wrap(~am)
p

# Major gridlines by default
p + theme_tntp()
# Without gridlines
p + theme_tntp(grid_color = "white")
}

