% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntp_palette.R
\name{tntp_palette}
\alias{tntp_palette}
\title{TNTP branded color palettes}
\usage{
tntp_palette(...)
}
\arguments{
\item{...}{supply quoted color names to include in color palette}
}
\description{
This function creates user defined color palettes using any of the
following nine TNTP colors: dark_blue, medium_blue, light_blue, green, orange,
gold, dark_grey (dark_gray), medium_grey (medium_gray), light_grey
(light_gray).
}
\examples{
library(ggplot2)

tntp_pal1 <- tntp_palette("green", "gold", "orange")
tntp_pal2 <- tntp_palette("dark_blue", "medium_blue", "light_blue")

p <- ggplot(mtcars, aes(wt, mpg))
p <- p + geom_point(aes(colour = factor(cyl)))
p

# Change colors to created palette
p <- p + scale_color_manual(values = tntp_pal1)
p

g <- ggplot(mtcars, aes(factor(cyl), mean(mpg)))
g <- g + geom_bar(aes(fill = factor(cyl)), stat = "identity")
g

# Change fill to created palette
g <- g + scale_fill_manual(values = tntp_pal2)
g
}

