% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntp_style.R
\name{tntp_style}
\alias{tntp_style}
\title{Create TNTP themed \link{ggplot2} charts}
\usage{
tntp_style(
  font = "Segoe UI",
  base_size = 28,
  text_color = "#222222",
  caption_color = "#7D7E81",
  show_legend_title = FALSE,
  show_axis_titles = FALSE,
  grid = FALSE,
  grid_color = "#CBCBCB",
  title_align = "left",
  legend_align = "left",
  caption_align = "right"
)
}
\arguments{
\item{font}{base font family}

\item{base_size}{base font size. Recommended minimum value of 15.}

\item{text_color}{text color for titles, axes, legends, and facets}

\item{caption_color}{text color for caption}

\item{show_legend_title}{Logical. Should the legend title be shown?  Leave as \code{TRUE} if you want to change the legend title with a subsequent line \code{+ labs(...)}.}

\item{show_axis_titles}{Which axis titles should be shown? Use \code{TRUE} or \code{FALSE} for toggle both titles, or \code{x} or \code{y} to show just that axis title.}

\item{grid}{Which grid lines should be shown? Use \code{TRUE} or \code{FALSE} to toggle all grid lines, or a string combination of \code{X}, \code{x}, \code{Y}, \code{y} for major and minor x and y grid lines.}

\item{grid_color}{grid line color}

\item{title_align, legend_align, caption_align}{Alignment of title, legend, and caption. Accepts \code{left}, \code{right}, or \code{center}}
}
\description{
A custom theme including TNTP fonts and other defaults for styling ggplot2 charts.
}
\examples{
library(ggplot2)
library(dplyr)

# Bar Chart with default settings
count(mpg, class) |>
  ggplot(aes(class, n)) +
  geom_col() +
  geom_text(aes(label = n), nudge_y = 3) +
  labs(x = "", y = "",
       title = "Car Types in the mpg Dataset",
       subtitle = "A plot that is only useful for demonstration purposes",
       caption = "Brought to you by the letter 'g'") +
  tntp_style()

# Default save size recommendation width = 9, height = 6:
# ggsave("filename.svg", width = 9, height = 6)

# Scatterplot including some customized settings
ggplot(mpg, aes(displ, hwy, color = class)) +
  geom_point(size = 3) +
  labs(x = "Engine Displacement", y = "MPG", color = "Class:",
       title = "Fuel Efficiency by Engine Size",
       subtitle = "An in-depth investigation with a very long subtitle that won't fit with default sizing",
       caption = "Brought to you by the letter 'g'") +
  scale_y_continuous(limits = c(0, 50)) +
  geom_hline(yintercept = 0, linewidth = 1, color = "#333333") +
  tntp_style(show_axis_titles = TRUE,
             show_legend_title = TRUE,
             legend_align = "center",
             grid = 'Yy',
             base_size = 20,
             text_color = tntpr::palette_tntp("dark_blue")) +
  facet_wrap(~drv)

}
