% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tntpr.R
\name{update_tntpr}
\alias{update_tntpr}
\title{Re-install the tntpr package from GitHub.}
\usage{
update_tntpr()
}
\value{
nothing
}
\description{
Re-install the tntpr package from GitHub.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Run without loading tntpr first
tntpr::update_tntpr()
\dontshow{\}) # examplesIf}
}
