% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart_counts.R
\name{bar_chart_counts}
\alias{bar_chart_counts}
\title{Bar chart of counts with TNTP polish}
\usage{
bar_chart_counts(df = NULL, var, group_var, label = "n",
  var_color = "medium_blue", group_colors, title = NULL, var_label,
  digits = 1, font = "Segoe UI", font_size = 12)
}
\arguments{
\item{df}{the data.frame to be used in the bar chart}

\item{var}{unquoted column name for variable to count}

\item{group_var}{(optional) unquoted column name for group variable.  If this is specified, you get a 2-variable clustered bar chart.  If left blank, a single variable bar chart.}

\item{label}{should labels show the count or the percentage?}

\item{var_color}{color for non-grouped charts; set to medium_blue by default}

\item{group_colors}{character vector of group colors, if a specific palette is desired}

\item{title}{main chart title}

\item{var_label}{label for x-axis}

\item{digits}{integer indicating the number of decimal places to be used in percentages}

\item{font}{font for chart text; Segoe UI by default}

\item{font_size}{size for chart text; set to 12 by default}
}
\description{
Takes a user supplied data frame and turns the designated column into
an N bar chart (uses position dodge from ggplot2).
}
\examples{

# An N bar chart by default
mtcars \%>\%
 bar_chart_counts(var     = cyl,
                  title   = "Number of mtcars by cylinder")

# Use a grouping variable
mtcars \%>\%
 bar_chart_counts(var          = cyl,
                  group_var    = vs,
                  label        = "pct",
                  title        = "Percentage of V vs. Straight engines by # of cylinders")

# Change default color
#'mtcars \%>\%
 bar_chart_counts(var       = cyl,
                  var_color = "orange"
                  title     = "Number of mtcars by cylinder")
# Specify color by group
 bar_chart_counts(mtcars, am, cyl, group_colors = c("orange", "green", "dark_blue"), label = "pct")
}

