% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart_tntp.R
\name{bar_chart_tntp}
\alias{bar_chart_tntp}
\title{Bar chart with TNTP polish}
\usage{
bar_chart_tntp(df = NULL, var, group_var, group_colors, title = NULL,
  var_label, font = "Segoe UI", font_size = 12,
  var_color = palette_tntp("dark_blue"), ...)
}
\arguments{
\item{df}{the data.frame to be used in the bar chart}

\item{var}{unquoted column name for desired variable}

\item{title}{main chart title}

\item{var_label}{label for x-axis}

\item{font}{font for chart text; Segoe UI by default}

\item{font_size}{size for chart text; set to 12 by default}

\item{var_color}{color for non-grouped charts; set to dark_blue by default}

\item{...}{additional arguments}

\item{(optional)}{group_var unquoted column name for group variable}

\item{(optional)}{group_colors character vector of group colors}

\item{display}{bar chart type "n" by default; also accepts "pct"}

\item{digits}{integer indicating the number of decimal places to be used}
}
\description{
Takes a user supplied data frame and turns the designated column into
an N bar chart (uses position dodge from ggplot2).
}
\examples{

# An  bar chart by default
mtcars \%>\%
 bar_chart_tntp(var     = cyl,
                title   = "Number of mtcars by cylinder")

# With a grouping variable
mtcars \%>\%
 bar_chart_tntp(var          = cyl,
                group_var    = vs,
                title        = "Percentage of mtcars by cylinder")
                
# with colors
bar_chart_tntp(mtcars, am, cyl, group_colors = palette_tntp("orange", "light_grey", "dark_blue"))
}

