% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart_tntp.R
\name{bar_chart_tntp}
\alias{bar_chart_tntp}
\title{Bar chart with TNTP polish}
\usage{
bar_chart_tntp(df = NULL, var, group_var, group_colors, title = NULL,
  font = "Segoe UI", font_size = 12,
  var_color = palette_tntp("light_grey"), ...)
}
\arguments{
\item{df}{the data.frame to be used in the bar chart}

\item{var}{unquoted column name for desired variable}

\item{group_var}{unquoted column name for group variable}

\item{group_colors}{character vector of group colors}

\item{title}{main chart title}

\item{font}{font for chart text; Segoe UI by default}

\item{font_size}{size for chart text; set to 12 by default}

\item{var_color}{color for non-grouped charts; set to light_grey by default}

\item{...}{additional arguments}
}
\description{
Takes a user supplied data frame and turns the designated column into
an N bar chart (uses position dodge from ggplot2).
}
\examples{

# An  bar chart by default
mtcars \%>\%
 bar_chart_tntp(var     = cyl,
                title   = "Number of mtcars by cylinder")

# With a grouping variable
mtcars \%>\%
 bar_chart_tntp(var          = cyl,
                group_var    = vs,
                title        = "Percentage of mtcars by cylinder")
}

