% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntp_colors.R
\name{choose_text_color}
\alias{choose_text_color}
\title{Choose a text color given a background color}
\usage{
choose_text_color(bg_color)
}
\arguments{
\item{bg_color}{a color}
}
\value{
"black" or "white"
}
\description{
Choose a text color given a background color
}
