% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_tntp.R
\docType{data}
\name{colors_tntp}
\alias{colors_tntp}
\title{TNTP colors}
\format{An object of class \code{character} of length 40.}
\usage{
colors_tntp
}
\description{
TNTP colors
}
\examples{
\dontrun{
 colors_tntp
}
}
\keyword{datasets}
