% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_tntp.R
\docType{data}
\name{colors_tntp_likert}
\alias{colors_tntp_likert}
\title{Likert pallette}
\format{
An object of class \code{character} of length 7.
}
\usage{
colors_tntp_likert
}
\description{
This likert palette has been superseded by the new brand colors and the new
function \code{\link[=tntp_palette]{tntp_palette()}}.
}
\examples{
\dontrun{
tntp_palette('likert_6')
}
}
\keyword{datasets}
