% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_tntp.R
\docType{data}
\name{colors_tntp_likert_orange_to_green}
\alias{colors_tntp_likert_orange_to_green}
\title{Likert orange to green pallette}
\format{
An object of class \code{character} of length 7.
}
\usage{
colors_tntp_likert_orange_to_green
}
\description{
This likert palette has been superseded by the new brand colors and the new
functions \code{\link[=tntp_colors]{tntp_colors()}} and \code{\link[=tntp_palette]{tntp_palette()}}.
}
\examples{
tntp_palette('bg_6')

}
\keyword{datasets}
