% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_agreement.R
\name{factorize_agreement}
\alias{factorize_agreement}
\title{Title}
\usage{
factorize_agreement(x, lvls)
}
\arguments{
\item{x}{the vector to factorize}

\item{lvls}{(optional) the levels, in order, of the resulting factor.  These should be provided as title case (see ?tools::toTitleCase).}
}
\value{
the vector, now as an ordered factor.  Currently the export will always return title case.
}
\description{
Title
}
\examples{

x <- c("Strongly Agree", "Disagree")
janitor::tabyl(x) # not ordered correctly
y <- factorize_agreement(x)
janitor::tabyl(y) # ordered correctly, shows missing levels
}
