% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize.R
\name{factorize_df}
\alias{factorize_df}
\title{Convert all character vectors containing a set of values in a data.frame to factors.}
\usage{
factorize_df(dat, lvls)
}
\arguments{
\item{dat}{data.frame with some factor variables stored as characters.}

\item{lvls}{The factor levels in your variable(s), in order.  If you have a question whose possible responses are a subset of another question's, don't use this function; manipulate the specific columns with \code{dplyr::mutate_at}.}
}
\value{
data.frame with factorization completed in place.
}
\description{
This function examines each column in a data.frame; when it finds a column composed solely of the values provided to the \code{lvls} argument it updates them to be factor variables, with levels in the order provided.

This is an alternative to calling \code{dplyr::mutate_at} with \code{factor()} and identifying the specific variables you want to transform, if you have several repeated sets of responses.
}
\examples{
library(dplyr)
mtcars \%>\%
  dplyr::mutate(agr = rep(c("Somewhat agree", "Strongly disagree"), 16)) \%>\%
  factorize_df(lvls = c("Strongly disagree", "Somewhat disagree",
                        "Somewhat agree", "Strongly agree"))

# prints warning due to case mismatches:
mtcars \%>\%
  dplyr::mutate(agr = rep(c("Somewhat Agree", "Strongly Disagree"), 16)) \%>\%
  factorize_df(lvls = c("Strongly disagree", "Somewhat disagree",
                        "Somewhat agree", "Strongly agree"))
}
