% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-utils.R
\name{figureN}
\alias{figureN}
\title{Create sequential figure numbers}
\usage{
figureN(x)
}
\arguments{
\item{x}{character string description of the figure}
}
\value{
nothing
}
\description{
Create sequential figure numbers
}
\examples{
# not run, in RMarkdown doc: `r figureN("Distribution of cars by cylinder count")`

#
}
