% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_tntp.R
\name{header_tntp}
\alias{header_tntp}
\title{Insert header_script_tntp.}
\usage{
header_tntp()
}
\value{
nothing
}
\description{
Call this function inside a .R file to insert the standard TNTP header.
}
\examples{
\dontrun{
header_tntp()
}

}
