% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_tntp.R
\docType{data}
\name{palette_names}
\alias{palette_names}
\title{Pallette names}
\format{
An object of class \code{character} of length 7.
}
\usage{
palette_names
}
\description{
Pallette names
}
\examples{
\dontrun{
palette_names
}
}
\keyword{datasets}
