% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_factor.R
\name{safe_factor}
\alias{safe_factor}
\alias{safe_ordered}
\title{Set factor levels safely}
\usage{
safe_factor(x = character(), levels, ...)

safe_ordered(x = character(), levels, ...)
}
\arguments{
\item{x}{a vector of data, usually taking a small number of distinct values}

\item{levels}{a vector of unique values}

\item{...}{additional arguments passed on to factor() or ordered()}
}
\value{
a factor
}
\description{
The default behavior of `base::factor()` and `base::ordered()` will silently
coerce values to NA if they don't match the provided levels.

These functions call the base functions, but provide an informative error
if any values in the vector would be coerced.
}
\examples{

# No error:
tntpr::teacher_survey |>
  dplyr::mutate(timing = safe_factor(timing, levels = c("Pre", "Post")))

\donttest{
# Mis-typed level generates an error
tntpr::teacher_survey |>
  dplyr::mutate(timing = safe_factor(timing, levels = c("Pre", "Pst")))
}
}
