% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_tntp.R
\name{scale_colour_tntp}
\alias{scale_colour_tntp}
\alias{scale_color_tntp}
\title{scale_colour_tntp/scale_color_tntp}
\usage{
scale_colour_tntp(palette = c("default", "likert_4pt", "likert_5pt",
  "likert_6pt", "colors_tntp_old"), ...)

scale_color_tntp(palette = c("default", "likert_4pt", "likert_5pt",
  "likert_6pt", "colors_tntp_old"), ...)
}
\arguments{
\item{palette}{character string describing the desired palette}

\item{...}{other arguments to pass through to ggplot2::discrete_scale()}
}
\description{
scale_colour_tntp/scale_color_tntp
}
