% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_tntp.R
\name{scale_fill_tntp}
\alias{scale_fill_tntp}
\title{scale_fill_tntp}
\usage{
scale_fill_tntp(palette = palette_names, drop = FALSE, ...)
}
\arguments{
\item{palette}{One of: "default", "likert_4pt", "likert_5pt", "likert_6pt", "colors_tntp_old"}

\item{drop}{drop}

\item{...}{additional arguments passed to ggplot2::discrete_scale()}
}
\description{
This function is deprecated. Please use
\code{scale_fill_manual(values = tntp_palette(palette_name))} instead
}
\examples{
library(ggplot2)
library(dplyr)

x <- mtcars \%>\%
  count(cyl, am) \%>\%
  mutate(am = as.factor(am))

ggplot(x, aes(x = cyl, y = n, fill = am)) + # you need a fill aesthetic
  geom_col() +
  scale_fill_manual(values = tntp_palette())

}
