% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tntp_2018.R
\docType{data}
\name{font_se}
\alias{font_se}
\title{Segoe UI font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_se
}
\description{
\code{font_an} == "\verb{Segoe UI}"
}
\keyword{datasets}
