% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_repo.R
\name{setup_repo}
\alias{setup_repo}
\title{Initialize a new repository, and a single subfolder, TNTP style.}
\usage{
setup_repo(subfolder, proj_name, analyst_name)
}
\arguments{
\item{subfolder}{a character vector containing the concise name of a project subfolder.
E.g., if the repository is the name of a city "Anywhere City", a project subfolder might be
"ela_access" or "aps_talent_landscape").}

\item{proj_name}{the longer, full name of the subfolder project.  This will
appear in the subfolder's README.md file.  E.g., "Access to Grade-Level ELA Content Pilot."}

\item{analyst_name}{the name(s) of the analysts currently working on the subfolder
project.  This will appear in the subfolder's README.md file.}
}
\description{
Create a new repository on Bitbucket, then set your working
directory to that folder and run this function.  It will set up the main
repo folder as well as a single subfolder in which you can work on your
immediate project.

You must specify the subfolder name as well as the long name associated with
that project and the analyst(s) working on it.  These latter two values are
used to create a README.Md file.
}
\examples{
# After the user has created a repository "Anywhere City" and set their working
# directory to that folder:
## not run
# setup_repo("ela_access",
# "Access to Grade-Level ELA Content",
# "Dustin Pashouwer and Sam Firke")
}
