% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_in_excel.R
\name{show_in_excel}
\alias{show_in_excel}
\title{Write datatable to temp excel file and open it.}
\usage{
show_in_excel(.data)
}
\arguments{
\item{.data}{Dataframe}
}
\value{
Returns a data.frame.
}
\description{
Write datatable to temp excel file and open it.
}
