% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_.R
\name{sp_error}
\alias{sp_error}
\title{For parsing sharepoint errors (right now just path errors)}
\usage{
sp_error(cnd, path)
}
\arguments{
\item{cnd}{condition}

\item{path}{path}
}
\value{
nothing
}
\description{
For parsing sharepoint errors (right now just path errors)
}
