% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_.R
\name{sp_list}
\alias{sp_list}
\alias{sp_list_drives}
\title{List Sharepoint Contents}
\usage{
sp_list(
  folder = "",
  site = NULL,
  drive = NULL,
  pattern = NULL,
  full_names = FALSE,
  recursive = FALSE,
  include_dirs = FALSE
)

sp_list_drives(site = NULL, pattern = NULL)
}
\arguments{
\item{folder}{Path to the folder. By default, lists the top-level contents of the drive.}

\item{site}{Site identifier. Can be the site_name, site_url or site_id. If not provided, uses the stored default site if it exists.}

\item{drive}{Name of the drive within the site. If site is provided but drive is not, uses the first drive of the provided site. If neither is provided, uses the stored default drive if it exists.}

\item{pattern}{Optional regular expression. Only names which match the regular expression will be returned.}

\item{full_names}{logical. If TRUE, the directory path is prepended to the file names to give a relative file path. If FALSE, the file names (rather than paths) are returned.}

\item{recursive}{logical. Should the listing recurse into directories? If TRUE, full_names is also set to TRUE.}

\item{include_dirs}{logical. Should subdirectory names be included in recursive listings? (They always are in non-recursive ones)}
}
\value{
A tibble with name and additional information on the relevant drives/files
}
\description{
`sp_list()` lists the contents of a Sharepoint Drive or a folder.
}
\details{
`sp_list_drives()` lists the drives contained in a Sharepoint site.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# List drives from the default site
sp_list_drives()

# List drives from a specific site
sp_list_drives("Data Analytics")
\dontshow{\}) # examplesIf}
}
