% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_.R
\name{sp_read}
\alias{sp_read}
\alias{sp_write}
\title{Read/Write from Sharepoint}
\usage{
sp_read(path, site = NULL, drive = NULL, type = NULL)

sp_write(x, path, site = NULL, drive = NULL, type = NULL)
}
\arguments{
\item{path}{The location in the Sharepoint drive}

\item{site}{Site identifier. Can be the site_name, site_url or site_id. If not provided, uses the stored default site if it exists.}

\item{drive}{Name of the drive within the site. If site is provided but drive is not, uses the first drive of the provided site. If neither is provided, uses the stored default drive if it exists.}

\item{type}{Optional. One of "rdata", "rds", or "dataframe". Uses the file extension to determine type if not provided.}

\item{x}{The object to be written}
}
\value{
`sp_read()` returns an R object as specified by type
`sp_write()` returns x, invisibly
}
\description{
Read or write data to/from a Sharepoint drive. If site and/or drive aren't
specified, uses the currently saved default from `sp_defaults()` if it exists.
}
\details{
These functions will attempt to use the appropriate read/write function based
on the file extension, however this can be overridden by specifying type
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Set site defaults
sp_defaults(site = "Data Analytics")

\dontshow{\}) # examplesIf}
}
