% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-utils.R
\name{tableN}
\alias{tableN}
\title{Create sequential table numbers}
\usage{
tableN(x)
}
\arguments{
\item{x}{character string description of the figure}
}
\value{
An atomic character vector prepended with a Table number
}
\description{
Create sequential table numbers
}
\examples{

tableN("Distribution of cars by cylinder count")
# Inline RMarkdown code: `r tableN("Distribution of cars by cylinder count")`
}
