% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tntp.R
\name{theme_tntp}
\alias{theme_tntp}
\title{TNTP's ggplot2 theme}
\usage{
theme_tntp(base_size = 12, base_family = "Segoe UI",
  grid_color = "grey93")
}
\arguments{
\item{base_size}{base font size; set to 12 by default}

\item{base_family}{base font family; set to "Segoe UI" by default}

\item{grid_color}{color for major gridlines; set to "grey93" by default}
}
\description{
ggplot2 theme customized for TNTP graphics
}
\examples{
p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg,
     colour=factor(gear))) + facet_wrap(~am)
p
p + theme_tntp()

# Without gridlines
p + theme_tntp(grid_color = "white")
}

