% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tntp.R
\name{theme_tntp}
\alias{theme_tntp}
\title{TNTP's ggplot2 theme}
\usage{
theme_tntp(
  show_legend_title = TRUE,
  base_size = 12,
  base_family = "Segoe UI",
  grid_color = "grey93",
  title_align = "center",
  title_color = "black",
  title_size = 12,
  subtitle_align = "center",
  subtitle_color = "black",
  subtitle_size = 12,
  caption_align = "right",
  caption_color = "black",
  caption_size = 12
)
}
\arguments{
\item{show_legend_title}{logical. Should the legend title be shown?  Leave as \code{TRUE} if you want to change the legend title with a subsequent line \code{+ labs(...)}.}

\item{base_size}{base font size}

\item{base_family}{base font family}

\item{grid_color}{color for major gridlines}

\item{title_align}{alignment of main title, defaults to "center"; also accepts "left" or "right"}

\item{title_color}{color of title text}

\item{title_size}{size of title text}

\item{subtitle_align}{alignment of sub-title, defaults to "center"; also accepts "left" or "right"}

\item{subtitle_color}{color of subtitle text}

\item{subtitle_size}{size of subtitle text}

\item{caption_align}{alignment of caption, defaults to "right"; also accepts "left" or "center"}

\item{caption_color}{color of caption text}

\item{caption_size}{size of caption text}
}
\description{
ggplot2 theme customized for TNTP aesthetics
}
