% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntpr.R
\docType{package}
\name{tntpr}
\alias{tntpr}
\alias{tntpr-package}
\title{Data Analysis Tools Customized for TNTP}
\description{
An in-house TNTP R package. Includes tools for data manipulation,
analysis, and reporting, including making TNTP-themed charts and documents.
By and for TNTP data-using staff, though available to the broader public.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tntp/tntpr}
  \item \url{https://tntp.github.io/tntpr/}
}

}
\author{
\strong{Maintainer}: Dustin Pashouwer \email{dustin.pashouwer@tntp.org}

Authors:
\itemize{
  \item Sam Firke
  \item Shane Orr \email{shane.orr@tntp.org}
  \item Sam Talcott \email{sam.talcott@tntp.org}
}

}
