% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tntp_2018.R
\name{update_geom_font_defaults}
\alias{update_geom_font_defaults}
\title{Update matching font defaults for text geoms}
\usage{
update_geom_font_defaults(
  family = "Segoe UI",
  face = "plain",
  size = 3.5,
  color = "#2b2b2b"
)
}
\arguments{
\item{family, face, size, color}{font family name, face, size and color}
}
\description{
Updates [ggplot2::geom_label] and [ggplot2::geom_text] font defaults
}
